refactor(survey): do not use static controller method in ui
